Extract and fix cw721 revoke tests
