+++ src/lib.rs
@@ -11,6 +11,7 @@ mod unittest_batch_mint;
 mod unittest_handles;
 mod unittest_burn;
 mod unittest_cw721_approve;
+mod unittest_cw721_revoke;
 mod unittest_helpers;
 mod unittest_mint;
 mod unittest_private_metadata;
